Struct casper_dao_erc721::ERC721Token

pub struct ERC721Token { /* private fields */ }

A module implementing ERC721 standard interface.

Implementations

impl ERC721Token

pub fn owner_of(&self, token_id: TokenId) -> Option<Address>

Finds the owner of a token with the given id.

Errors

Reverts with Error::TokenDoesNotExist if token with the given id does not exists.

pub fn balance_of(&self, owner: Address) -> U512

Counts all tokens assigned to an owner.

pub fn total_supply(&self) -> U512

Return the total number of tokens.

pub fn approve(&mut self, approved: Option<Address>, token_id: TokenId)

Sets or revokes the approved address for a token.

Errors

Reverts with Error::TokenDoesNotExist if token with the given id does not exists. Reverts with Error::ApprovalToCurrentOwner to the approved address is the owner. Reverts with Error::ApproveCallerIsNotOwnerNorApprovedForAll if the caller is neither the owner not approved for all. See ERC721Token::set_approval_for_all().

pub fn get_approved(&self, token_id: TokenId) -> Option<Address>

Gets the approved address for a token.

pub fn set_approval_for_all(&mut self, operator: Address, approved: bool)

Enables or disables approval for a third party (“operator”) to manage all of caller’s tokens

Errors

Reverts with Error::ApproveToCaller the caller tries to approve himself.

pub fn is_approved_for_all(&self, owner: Address, operator: Address) -> bool

Checks if an address is an authorized operator for another address

pub fn transfer_from(
&mut self,
owner: Address,
recipient: Address,
token_id: TokenId
)

Transfers the ownership of a token from one address to another address.

Errors

The caller must be an owner or be an approved address, otherwise the contract reverts with Error::CallerIsNotOwnerNorApproved.

pub fn safe_transfer_from(
&mut self,
owner: Address,
recipient: Address,
token_id: TokenId,
data: Option<Bytes>
)

Transfers the ownership of a token from one address to another address.

Verifies whether the recipient is a smart contract that implements ERC721Receiver.

Errors

The caller must be an owner or be an approved address, otherwise the contract reverts with Error::CallerIsNotOwnerNorApproved.

impl ERC721Token

pub fn owner_of_or_revert(&self, token_id: TokenId) -> Address

pub fn increment_balance(&mut self, owner: Address)

pub fn decrement_balance(&mut self, owner: Address)

pub fn increment_total_supply(&mut self)

pub fn decrement_total_supply(&mut self)

pub fn exists(&self, token_id: &TokenId) -> bool

pub fn set_owner_of(&mut self, token_id: TokenId, owner: Option<Address>)

pub fn approve_owner(
&mut self,
owner: Option<Address>,
approved: Option<Address>,
token_id: TokenId
)